+++ google/src/main/scala/org/broadinstitute/dsde/workbench/google/GoogleKmsInterpreter.scala
@@ -6,10 +6,10 @@ import com.google.api.gax.core.FixedCredentialsProvider
 import com.google.auth.oauth2.ServiceAccountCredentials
 import com.google.cloud.kms.v1.CryptoKey.CryptoKeyPurpose
 import com.google.cloud.kms.v1._
+import com.google.common.util.concurrent.ThreadFactoryBuilder
 import com.google.iam.v1.{Binding, Policy}
 import com.google.protobuf.{Duration, Timestamp}
 import org.broadinstitute.dsde.workbench.model.google.GoogleProject
-
 import scala.jdk.CollectionConverters._
 import scala.language.higherKinds
 
@@ -159,7 +159,14 @@ object GoogleKmsInterpreter {
   def apply[F[_]: Sync](client: KeyManagementServiceClient): GoogleKmsInterpreter[F] =
     new GoogleKmsInterpreter[F](client)
 
-  def client[F[_]: Sync](pathToJson: String): Resource[F, KeyManagementServiceClient] =
+  def client[F[_]: Sync](pathToJson: String): Resource[F, KeyManagementServiceClient] = {
+    val executorProviderBuilder = KeyManagementServiceSettings.defaultExecutorProviderBuilder()
+    val threadFactory = new ThreadFactoryBuilder()
+      .setThreadFactory(executorProviderBuilder.getThreadFactory)
+      .setNameFormat("goog-kms-%d")
+      .build()
+    val executorProvider = executorProviderBuilder.setThreadFactory(threadFactory).build()
+
     for {
       credentials <- org.broadinstitute.dsde.workbench.util2.readFile(pathToJson)
       client <- Resource.make[F, KeyManagementServiceClient](
@@ -167,6 +174,7 @@ object GoogleKmsInterpreter {
           KeyManagementServiceClient.create(
             KeyManagementServiceSettings
               .newBuilder()
+              .setBackgroundExecutorProvider(executorProvider)
               .setCredentialsProvider(
                 FixedCredentialsProvider.create(ServiceAccountCredentials.fromStream(credentials))
               )
@@ -175,4 +183,5 @@ object GoogleKmsInterpreter {
         )
       )(client => Sync[F].delay(client.close()))
     } yield client
+  }
 }